Improve ManageTagController coverage and error handling
